+++ stories/index.tsx
@@ -20,6 +20,7 @@ import '../src/layout/section/tripsSection/story'
 import '../src/layout/section/mediaContentSection/story'
 import '../src/layout/section/columnedContentSection/story'
 import '../src/layout/section/illustratedSection/story'
+import '../src/layout/section/itemsSection/story'
 
 // Pages
 import '../src/pages/messaging/inbox.story'